+++ src/main/kotlin/com/github/hoxo/fileservice/dto/FileChunkDto.kt
@@ -3,7 +3,7 @@ package com.github.hoxo.fileservice.dto
 data class FileChunkDto(
     val info: FileInfoDto,
     val offset: Long,
-    val size: Long,
+    val size: Int,
     val data: String,
     val hasRemainingData: Boolean,
 )
